+++ Javascripts/portfolios/loadPortfolios.js
@@ -1,153 +1,7 @@
 import { portfolioClicked } from "../navBtnSelector.js";
-
-const main = document.querySelector('main');
-const styleDiv = document.getElementById('stylesheetContainer');
+import { portfoliosHtmlLoad } from "./portfolios_html_loader.js";
 
 export function loadPortfolios() {
-    styleDiv.innerHTML = `<link rel="stylesheet" href="./Stylesheets/portfolios.css">`;
+    portfoliosHtmlLoad();
     portfolioClicked();
-    main.innerHTML = `<a class="card">
-    <div class="content">
-        <div class="front flexbox">
-            <h3 class="title">Hey</h3>
-        </div>
-        <div class="back flexbox">
-            <p class="description">
-                This is a test
-            </p>
-        </div>
-    </div>
-</a>
-<a class="card" href="https://www.youtube.com/watch?v=dQw4w9WgXcQ&ab_channel=RickAstley" target="_blank">
-    <div class="content">
-        <div class="front flexbox">
-            <h3 class="title">Hey</h3>
-        </div>
-        <div class="back flexbox">
-            <p class="description">
-                This is a test
-            </p>
-        </div>
-    </div>
-</a>
-<a class="card" href="https://www.youtube.com/watch?v=dQw4w9WgXcQ&ab_channel=RickAstley" target="_blank">
-    <div class="content">
-        <div class="front flexbox">
-            <h3 class="title">Hey</h3>
-        </div>
-        <div class="back flexbox">
-            <p class="description">
-                This is a test
-            </p>
-        </div>
-    </div>
-</a>
-<a class="card" href="https://www.youtube.com/watch?v=dQw4w9WgXcQ&ab_channel=RickAstley" target="_blank">
-    <div class="content">
-        <div class="front flexbox">
-            <h3 class="title">Hey</h3>
-        </div>
-        <div class="back flexbox">
-            <p class="description">
-                This is a test
-            </p>
-        </div>
-    </div>
-</a>
-<a class="card" href="https://www.youtube.com/watch?v=dQw4w9WgXcQ&ab_channel=RickAstley" target="_blank">
-    <div class="content">
-        <div class="front flexbox">
-            <h3 class="title">Hey</h3>
-        </div>
-        <div class="back flexbox">
-            <p class="description">
-                This is a test
-            </p>
-        </div>
-    </div>
-</a>
-<a class="card" href="https://www.youtube.com/watch?v=dQw4w9WgXcQ&ab_channel=RickAstley" target="_blank">
-    <div class="content">
-        <div class="front flexbox">
-            <h3 class="title">Hey</h3>
-        </div>
-        <div class="back flexbox">
-            <p class="description">
-                This is a test
-            </p>
-        </div>
-    </div>
-</a>
-<a class="card">
-    <div class="content">
-        <div class="front flexbox">
-            <h3 class="title">Hey</h3>
-        </div>
-        <div class="back flexbox">
-            <p class="description">
-                This is a test
-            </p>
-        </div>
-    </div>
-</a>
-<a class="card" href="https://www.youtube.com/watch?v=dQw4w9WgXcQ&ab_channel=RickAstley" target="_blank">
-    <div class="content">
-        <div class="front flexbox">
-            <h3 class="title">Hey</h3>
-        </div>
-        <div class="back flexbox">
-            <p class="description">
-                This is a test
-            </p>
-        </div>
-    </div>
-</a>
-<a class="card" href="https://www.youtube.com/watch?v=dQw4w9WgXcQ&ab_channel=RickAstley" target="_blank">
-    <div class="content">
-        <div class="front flexbox">
-            <h3 class="title">Hey</h3>
-        </div>
-        <div class="back flexbox">
-            <p class="description">
-                This is a test
-            </p>
-        </div>
-    </div>
-</a>
-<a class="card" href="https://www.youtube.com/watch?v=dQw4w9WgXcQ&ab_channel=RickAstley" target="_blank">
-    <div class="content">
-        <div class="front flexbox">
-            <h3 class="title">Hey</h3>
-        </div>
-        <div class="back flexbox">
-            <p class="description">
-                This is a test
-            </p>
-        </div>
-    </div>
-</a>
-<a class="card" href="https://www.youtube.com/watch?v=dQw4w9WgXcQ&ab_channel=RickAstley" target="_blank">
-    <div class="content">
-        <div class="front flexbox">
-            <h3 class="title">Hey</h3>
-        </div>
-        <div class="back flexbox">
-            <p class="description">
-                This is a test
-            </p>
-        </div>
-    </div>
-</a>
-<a class="card" href="https://www.youtube.com/watch?v=dQw4w9WgXcQ&ab_channel=RickAstley" target="_blank">
-    <div class="content">
-        <div class="front flexbox">
-            <h3 class="title">Hey</h3>
-        </div>
-        <div class="back flexbox">
-            <p class="description">
-                This is a test
-            </p>
-        </div>
-    </div>
-</a>`;
 }
