+++ Javascripts/portfolios/portfolios_html_loader.js
@@ -1,0 +1,150 @@
+const main = document.querySelector('main');
+const styleDiv = document.getElementById('stylesheetContainer');
+
+export function portfoliosHtmlLoad(){
+    styleDiv.innerHTML = `<link rel="stylesheet" href="./Stylesheets/portfolios.css">`;
+    main.innerHTML = `<a class="card">
+    <div class="content">
+        <div class="front flexbox">
+            <h3 class="title">Hey</h3>
+        </div>
+        <div class="back flexbox">
+            <p class="description">
+                This is a test
+            </p>
+        </div>
+    </div>
+</a>
+<a class="card" href="https://www.youtube.com/watch?v=dQw4w9WgXcQ&ab_channel=RickAstley" target="_blank">
+    <div class="content">
+        <div class="front flexbox">
+            <h3 class="title">Hey</h3>
+        </div>
+        <div class="back flexbox">
+            <p class="description">
+                This is a test
+            </p>
+        </div>
+    </div>
+</a>
+<a class="card" href="https://www.youtube.com/watch?v=dQw4w9WgXcQ&ab_channel=RickAstley" target="_blank">
+    <div class="content">
+        <div class="front flexbox">
+            <h3 class="title">Hey</h3>
+        </div>
+        <div class="back flexbox">
+            <p class="description">
+                This is a test
+            </p>
+        </div>
+    </div>
+</a>
+<a class="card" href="https://www.youtube.com/watch?v=dQw4w9WgXcQ&ab_channel=RickAstley" target="_blank">
+    <div class="content">
+        <div class="front flexbox">
+            <h3 class="title">Hey</h3>
+        </div>
+        <div class="back flexbox">
+            <p class="description">
+                This is a test
+            </p>
+        </div>
+    </div>
+</a>
+<a class="card" href="https://www.youtube.com/watch?v=dQw4w9WgXcQ&ab_channel=RickAstley" target="_blank">
+    <div class="content">
+        <div class="front flexbox">
+            <h3 class="title">Hey</h3>
+        </div>
+        <div class="back flexbox">
+            <p class="description">
+                This is a test
+            </p>
+        </div>
+    </div>
+</a>
+<a class="card" href="https://www.youtube.com/watch?v=dQw4w9WgXcQ&ab_channel=RickAstley" target="_blank">
+    <div class="content">
+        <div class="front flexbox">
+            <h3 class="title">Hey</h3>
+        </div>
+        <div class="back flexbox">
+            <p class="description">
+                This is a test
+            </p>
+        </div>
+    </div>
+</a>
+<a class="card">
+    <div class="content">
+        <div class="front flexbox">
+            <h3 class="title">Hey</h3>
+        </div>
+        <div class="back flexbox">
+            <p class="description">
+                This is a test
+            </p>
+        </div>
+    </div>
+</a>
+<a class="card" href="https://www.youtube.com/watch?v=dQw4w9WgXcQ&ab_channel=RickAstley" target="_blank">
+    <div class="content">
+        <div class="front flexbox">
+            <h3 class="title">Hey</h3>
+        </div>
+        <div class="back flexbox">
+            <p class="description">
+                This is a test
+            </p>
+        </div>
+    </div>
+</a>
+<a class="card" href="https://www.youtube.com/watch?v=dQw4w9WgXcQ&ab_channel=RickAstley" target="_blank">
+    <div class="content">
+        <div class="front flexbox">
+            <h3 class="title">Hey</h3>
+        </div>
+        <div class="back flexbox">
+            <p class="description">
+                This is a test
+            </p>
+        </div>
+    </div>
+</a>
+<a class="card" href="https://www.youtube.com/watch?v=dQw4w9WgXcQ&ab_channel=RickAstley" target="_blank">
+    <div class="content">
+        <div class="front flexbox">
+            <h3 class="title">Hey</h3>
+        </div>
+        <div class="back flexbox">
+            <p class="description">
+                This is a test
+            </p>
+        </div>
+    </div>
+</a>
+<a class="card" href="https://www.youtube.com/watch?v=dQw4w9WgXcQ&ab_channel=RickAstley" target="_blank">
+    <div class="content">
+        <div class="front flexbox">
+            <h3 class="title">Hey</h3>
+        </div>
+        <div class="back flexbox">
+            <p class="description">
+                This is a test
+            </p>
+        </div>
+    </div>
+</a>
+<a class="card" href="https://www.youtube.com/watch?v=dQw4w9WgXcQ&ab_channel=RickAstley" target="_blank">
+    <div class="content">
+        <div class="front flexbox">
+            <h3 class="title">Hey</h3>
+        </div>
+        <div class="back flexbox">
+            <p class="description">
+                This is a test
+            </p>
+        </div>
+    </div>
+</a>`;
+}
